requirements_test.txt
pytest
pytest-asyncio
homeassistant
python-snap7
